Add files via upload
again, i fucked up and forgot to connect repo via vs, will fix tomorrow, if I am not grinding my soul to oblviion

once you login in, you are redirected to the home page. Once you click on the "Get started", you will be routed to "features" so everybody knows how to succeed.

profile page is now added. It is on the nav bar. Is based on what you did earlier, backend is added as support so its not just a floating html/view document. 

calendar is added. click "assement remidner" and it will redirect you there. still uses the hover effect of the other non-clickable button categories. Title, course, due date is in ISO 8061 (I think, I forgot), you can scroll down which date you prefer to set reminders for, I guess we're all from back to the future with these features, because what If i made a reminder like 10 years ago, how am i supposed to do that when I AM NOW IN THE FUTURE OMG

reminders are editable, they are deletable, and they are able to be created. enjoy.